user directory reorganization
