+++ app/config/services.php
@@ -10,13 +10,13 @@ use App\Article\Infrastructure\Presenter\JsonArticleListPresenter;
 use App\Article\Infrastructure\Presenter\JsonArticlePresenter;
 use App\Article\Infrastructure\Repository\ArticleRepository;
 use App\User\Application\Services\PasswordService;
-use App\User\Application\UseCase\LoginPresenterInterface;
-use App\User\Application\UseCase\LoginUserUseCase;
-use App\User\Application\UseCase\LoginUserUseCaseInterface;
-use App\User\Domain\Repository\UserRepository;
+use App\User\Application\UseCase\Login\LoginUserUseCase;
+use App\User\Application\UseCase\Login\LoginUserUseCaseInterface;
 use App\User\Domain\Repository\UserRepositoryInterface;
 use App\User\Infrastructure\Controller\LoginController;
+use App\User\Infrastructure\LoginPresenterInterface;
 use App\User\Infrastructure\Presenter\LoginPresenter;
+use App\User\Infrastructure\Repository\UserRepository;
 use Config\Config;
 use Szabek\Framework\Container;
 use Szabek\Framework\Http\Middleware\AuthMiddleware;
